+++ NEWS.md
@@ -1,3 +1,7 @@
+# November 2022
+* Releasing iotarelr 0.1.2
+* Providing Iota Concept of generation two
+
 # 29 March 2022
 * Releasing iotarelr 0.1.1
 * Providing Iota Concept of generation one